Experiments sync

+++ ptls/data_load/data_module/sop_data_module.py
@@ -48,7 +48,7 @@ def collate_sop_pairs(batch):
             padded_collate_wo_target(lefts),
             padded_collate_wo_target(rights)
         ),
-        targets.float()
+        targets.long()
     )
