+++ ptls/frames/bert/datasets/sop_dataset.py
@@ -26,7 +26,7 @@ class SopDataset(ColesDataset):
                 collate_feature_dict(lefts),
                 collate_feature_dict(rights)
             ),
-            targets.float()
+            targets.long()
         )
